Remove LogMetadata
